added stuff

+++ src/mesa/component/textbox.py
@@ -6,14 +6,12 @@ import pygame as pg
 
 
 class MesaTextBoxInput(_MesaContainer):
-    """
-    texttexttexttexttext
-    """
     def __init__(self, parent) -> None:
         super().__init__(parent)
         self.font_name = MesaCoreFlag.NOT_DECLARED_ON_INIT
         self.font = MesaCoreFlag.NOT_DECLARED_ON_INIT
         self.font_size = MesaCoreFlag.NOT_DECLARED_ON_INIT
+        self.font_type = "SYS"
         self.text = ""
         self.text_surface = MesaCoreFlag.NOT_DECLARED_ON_INIT
         self.bold = False
@@ -25,51 +23,80 @@ class MesaTextBoxInput(_MesaContainer):
         self.text_center_v_flag = MesaCoreFlag.NOT_DECLARED_ON_INIT
         self.text_center_h_flag = MesaCoreFlag.NOT_DECLARED_ON_INIT
         self.metrics = []
-        self.pointer_position = self.get_pointer_position()
+        self.pointer_position = self._get_pointer_position()
+        self.active = False
         self.blink = False
         self.tick = 0
+        self.offset = pg.Vector2(10, 0)
 
-    def get_input(self):
+    def declare_font_type(self, _type: str):
+        # This is subject to change, strings are considered bad practice.
+        if _type == "SYS":
+            self.font_type = "SYS"
+        if _type == "NOSYS":
+            self.font_type = "NOSYS"
+
+    def load_ttf(self, ttf):
+        self.font_name = ttf
+
+    def center_text_vertical(self):
+        self.text_center_v_flag = MesaRenderFlag.TEXT_CENTERED_V
+
+    def center_text_horizontal(self):
+        self.text_center_h_flag = MesaRenderFlag.TEXT_CENTERED_H
+
+    def center_text(self):
+        self.center_text_horizontal()
+        self.center_text_vertical()
+
+    def get_written_text(self):
         return self.buffer
 
-    def get_pointer_position(self):
+    def _get_pointer_position(self):
         return (sum([x[4] for x in self.metrics[: self.buffer.pointer]]),)
 
-    def handle_events(self):
+    def _handle_events(self):
         for event in self.scene.manager.get_events():
-            if event.type == pg.TEXTINPUT:
+            if event.type == pg.TEXTINPUT and self.active:
                 self.buffer.add(event.text)
                 self.text = self.buffer.buffer
-                self.make_text_surface()
+                self._make_text_surface()
                 self.buffer.shift_right()
                 self.metrics = pg.Font.metrics(self.font, self.text)
-                self.pointer_position = self.get_pointer_position()
+                self.pointer_position = self._get_pointer_position()
 
-            if event.type == pg.KEYDOWN:
+            if event.type == pg.KEYDOWN and self.active:
                 if event.key == pg.K_TAB:
                     self.buffer.delete()
                     self.text = self.buffer.buffer
-                    self.make_text_surface()
+                    self._make_text_surface()
                     self.metrics = pg.Font.metrics(self.font, self.text)
-                    self.pointer_position = self.get_pointer_position()
                 if event.key == 8:
                     self.buffer.pop()
                     self.text = self.buffer.buffer
-                    self.make_text_surface()
+                    self._make_text_surface()
                     self.metrics = pg.Font.metrics(self.font, self.text)
-                    self.pointer_position = self.get_pointer_position()
                 if event.key == pg.K_LEFT:
                     self.buffer.shift_left()
-                    self.pointer_position = self.get_pointer_position()
                 if event.key == pg.K_RIGHT:
                     self.buffer.shift_right()
-
-                    self.pointer_position = self.get_pointer_position()
                 if event.key == pg.K_RETURN:
                     self.text += "\n"
-                    self.make_text_surface()
+                    self._make_text_surface()
                     self.metrics = pg.Font.metrics(self.font, self.text)
-                    self.pointer_position = self.get_pointer_position()
+                    self.active = False
+                    pg.key.stop_text_input()
+                self.pointer_position = self._get_pointer_position()
+
+            if event.type == pg.MOUSEBUTTONDOWN:
+                if self.active:
+                    self.active = False
+                if self._is_container_hovered():
+                    if self.active == False:
+                        pg.key.start_text_input()
+                    else:
+                        pg.key.stop_text_input()
+                    self.active = not self.active
 
     def set_text_color(self, text_color):
         if self.text_color == MesaCoreFlag.NOT_DECLARED_ON_INIT:
@@ -77,7 +104,7 @@ class MesaTextBoxInput(_MesaContainer):
         if self.text_surface != MesaCoreFlag.NOT_DECLARED_ON_INIT:
             if text_color != self.text_color:
                 self.text_color = text_color
-                self.make_text_surface()
+                self._make_text_surface()
 
     def set_text_background_color(self, color):
         self.text_background_color = color
@@ -100,10 +127,13 @@ class MesaTextBoxInput(_MesaContainer):
         if self.font_size != MesaCoreFlag.NOT_DECLARED_ON_INIT:
             if font_size != self.font_size:
                 self.font_size = font_size
-                self.font = pg.font.SysFont(
-                    self.font_name, self.font_size, self.bold, self.italic
-                )
-                self.make_text_surface()
+                if self.font_type == "SYS":
+                    self.font = pg.font.SysFont(
+                        self.font_name, self.font_size, self.bold, self.italic
+                    )
+                elif self.font_type == "NOSYS":
+                    self.font = pg.font.Font(self.font_name, self.font_size)
+                self._make_text_surface()
 
     def set_text(self, text):
         if self.text == MesaCoreFlag.NOT_DECLARED_ON_INIT:
@@ -111,10 +141,9 @@ class MesaTextBoxInput(_MesaContainer):
         if self.text_surface != MesaCoreFlag.NOT_DECLARED_ON_INIT:
             if text != self.text:
                 self.text = text
-                self.make_text_surface()
+                self._make_text_surface()
 
-    def make_text_surface(self):
-        "called"
+    def _make_text_surface(self):
         self.text_surface = self.font.render(
             self.text,
             self.antialias,
@@ -123,39 +152,81 @@ class MesaTextBoxInput(_MesaContainer):
         )
 
     def late_init(self):
-        self.font = pg.font.SysFont(
-            self.font_name, self.font_size, self.bold, self.italic
-        )
-        self.make_text_surface()
+        if self.font_type == "SYS":
+            self.font = pg.font.SysFont(
+                self.font_name, self.font_size, self.bold, self.italic
+            )
+        elif self.font_type == "NOSYS":
+            self.font = pg.font.Font(self.font_name, self.font_size)
+        self._make_text_surface()
         pg.key.set_text_input_rect(self.rect)
         return super().late_init()
 
     def inherit_update(self):
         self.tick += 1
-        self.handle_events()
+        self._handle_events()
         return super().inherit_update()
 
-    def render(self):
+    def _configure_text_position(self):
         self.text_position = pg.Vector2(0, 0)
         if self.text_center_v_flag == MesaRenderFlag.TEXT_CENTERED_V:
             self.text_position.y = (self.height - self.text_surface.get_height()) // 2
-
         if self.text_center_h_flag == MesaRenderFlag.TEXT_CENTERED_H:
             self.text_position.x = (self.width - self.text_surface.get_width()) // 2
-        if self.tick % 5 == 0:
+
+    def _blink_logic(self):
+        if self.tick % 25 == 0:
             self.blink = not self.blink
+
+    def _draw_pointer(self):
+        overflow = pg.Vector2(
+            self.pointer_position[0] - self.surface.get_width() + self.offset.x + 30, 0
+        )
+        overflow.x = max(0, overflow.x)
+        blink_width = 10
+        blink_height = self.font_size + 2
+        if self.metrics:
+            blink_width = self.metrics[self.buffer.pointer - 1][4]
+
         if self.blink:
             pg.draw.rect(
                 self.surface,
-                "cyan",
-                [self.pointer_position[0], 0, 2, 24],
+                [100, 100, 100, 120],
+                [
+                    self.pointer_position[0] + self.offset.x - overflow.x,
+                    self.text_position.y,
+                    blink_width,
+                    blink_height,
+                ],
                 0,
             )
         else:
             pg.draw.rect(
                 self.surface,
-                "black",
-                [self.pointer_position[0], 0, 2, 24],
+                [100, 100, 250, 190],
+                [
+                    self.pointer_position[0] + self.offset.x - overflow.x,
+                    self.text_position.y,
+                    blink_width,
+                    blink_height,
+                ],
                 0,
             )
-        self.surface.blit(self.text_surface, self.text_position)
+
+    def _draw_text(self):
+        overflow = pg.Vector2(
+            self.pointer_position[0] - self.surface.get_width() + self.offset.x + 30, 0
+        )
+        if overflow.x > 0:
+            self.surface.blit(
+                self.text_surface, self.text_position + self.offset - overflow
+            )
+        else:
+            self.surface.blit(self.text_surface, self.text_position + self.offset)
+
+    def render(self):
+        self._configure_text_position()
+        if self.active:
+            self._blink_logic()
+            self._draw_pointer()
+        self._draw_text()
